# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smoothy King API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RecipeIngredient {
  name: string;
  amount: string;
  unit: string;
  /**
     * Primary skin/health benefit of this ingredient
     * @nullable
     */
  benefit?: string | null;
}

export interface Recipe {
  id: number;
  name: string;
  slug: string;
  /** smoothie | shaker | meal */
  category: string;
  /** Short poetic descriptor (e.g. 'The Glow Ritual') */
  tagline: string;
  /** @nullable */
  description?: string | null;
  prepTimeMinutes: number;
  servings: number;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  protein?: number | null;
  /** e.g. ['glowy-skin', 'hydration', 'anti-inflammatory'] */
  benefits: string[];
  /**
     * 1-10 skin benefit rating inspired by K-beauty principles
     * @nullable
     */
  skinBenefitScore?: number | null;
  ingredients: RecipeIngredient[];
  steps: string[];
  imageUrl: string;
  isFeatured: boolean;
  /**
     * easy | medium | advanced
     * @nullable
     */
  difficulty?: string | null;
  tags?: string[];
}

export interface Ingredient {
  id: number;
  name: string;
  /** fruit | vegetable | protein | adaptogen | superfood | fat | liquid */
  category: string;
  benefits: string[];
  /**
     * Main K-beauty inspired skin benefit
     * @nullable
     */
  skinBenefitKey?: string | null;
  description: string;
  /**
     * Korean name or beauty term if applicable
     * @nullable
     */
  koreanName?: string | null;
  nutrientHighlights?: string[];
}

export interface BenefitGroup {
  benefit: string;
  label: string;
  description: string;
  recipes: Recipe[];
}

export interface FavoriteInput {
  recipeId: number;
}

export interface Creation {
  id: number;
  name: string;
  authorName: string;
  /** @nullable */
  authorInitials?: string | null;
  /** The benefit goal this creation serves (e.g. glowy-skin) */
  goal: string;
  /**
     * Author's note about why they built this blend
     * @nullable
     */
  story?: string | null;
  ingredients: RecipeIngredient[];
  likes: number;
  /**
     * The blend's signature color for UI display
     * @nullable
     */
  colorHex?: string | null;
  createdAt: string;
}

export interface CreationInput {
  name: string;
  authorName: string;
  goal: string;
  story?: string;
  ingredients: RecipeIngredient[];
  colorHex?: string;
}

export interface Plan {
  id: number;
  name: string;
  tagline: string;
  pricePerMonth: number;
  features: string[];
  isPopular: boolean;
  /** @nullable */
  accentHex?: string | null;
}

export interface CommunityStats {
  members: number;
  creationsThisWeek: number;
  ritualsCompleted: number;
  topGoal: string;
}

export interface SignupInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  nickname: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserPublic {
  id: number;
  email: string;
  nickname: string;
  createdAt: string;
}

export interface AuthResponse {
  user: UserPublic;
  token: string;
}

export interface OnboardingInput {
  /** male | female | other */
  gender?: string;
  age?: number;
  /** cm */
  height?: number;
  /** kg */
  weight?: number;
  /** sedentary | light | moderate | active | very_active */
  activityLevel?: string;
  allergies?: string[];
  dislikedIngredients?: string[];
  /** glowy-skin | hydration | sun-ritual | protein-power | anti-inflammatory | detox-clarity */
  primaryGoal?: string;
  secondaryGoals?: string[];
  /** sweet | sour | nutty | fresh */
  tastePreference?: string[];
}

export interface UserProfileResponse {
  id: number;
  userId: number;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  activityLevel?: string | null;
  allergies?: string[];
  dislikedIngredients?: string[];
  /** @nullable */
  primaryGoal?: string | null;
  secondaryGoals?: string[];
  tastePreference?: string[];
  createdAt?: string;
  updatedAt?: string;
}

export type ListRecipesParams = {
category?: string;
benefit?: string;
search?: string;
};

export type ListCreationsParams = {
/**
 * recent | popular
 */
sort?: string;
goal?: string;
};