# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smoothy King API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          description: Not authenticated
...
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List all favorited recipe IDs
      responses:
        "200":
          description: Favorited recipe IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a recipe to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Added
  /favorites/{recipeId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a recipe from favorites
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /creations:
    get:
      operationId: listCreations
      tags: [creations]
      summary: List community smoothie creations
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
            description: "recent | popular"
        - name: goal
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Community creations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creation"
    post:
      operationId: createCreation
      tags: [creations]
      summary: Publish a custom smoothie creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creation"
  /creations/{id}/like:
    post:
      operationId: likeCreation
      tags: [creations]
      summary: Like a community creation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creation"
    delete:
      operationId: unlikeCreation
      tags: [creations]
      summary: Remove a like from a community creation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creation"
...
components:
  schemas:
...
    RecipeIngredient:
      type: object
      required: [name, amount, unit]
      properties:
        name:
          type: string
        amount:
          type: string
        unit:
          type: string
        benefit:
          type: ["string", "null"]
          description: "Primary skin/health benefit of this ingredient"
...
    FavoriteInput:
      type: object
      required: [recipeId]
      properties:
        recipeId:
          type: integer
    Creation:
      type: object
      required: [id, name, authorName, goal, ingredients, likes, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        authorName:
          type: string
        authorInitials:
          type: ["string", "null"]
        goal:
          type: string
          description: "The benefit goal this creation serves (e.g. glowy-skin)"
        story:
          type: ["string", "null"]
          description: "Author's note about why they built this blend"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        likes:
          type: integer
        colorHex:
          type: ["string", "null"]
          description: "The blend's signature color for UI display"
        createdAt:
          type: string
    CreationInput:
      type: object
      required: [name, authorName, goal, ingredients]
      properties:
        name:
          type: string
        authorName:
          type: string
        goal:
          type: string
        story:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        colorHex:
          type: string
...
    SignupInput:
      type: object
      required: [email, password, nickname]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        nickname:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
        token:
          type: string
    UserPublic:
      type: object
      required: [id, email, nickname, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        nickname:
          type: string
        createdAt:
          type: string
...