# lib/api-zod/src/generated/types/creationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smoothy King API
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeIngredient } from './recipeIngredient';

export interface CreationInput {
  name: string;
  authorName: string;
  goal: string;
  story?: string;
  ingredients: RecipeIngredient[];
  colorHex?: string;
}