# lib/api-zod/src/generated/types/recipeIngredient.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smoothy King API
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeIngredient {
  name: string;
  amount: string;
  unit: string;
  /**
     * Primary skin/health benefit of this ingredient
     * @nullable
     */
  benefit?: string | null;
}